% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit documents.R
\name{export_bundled_ggplot}
\alias{export_bundled_ggplot}
\title{Exports bundled ggplot object as it will appear in print}
\usage{
export_bundled_ggplot(bundled_ggplot, file, device = "png")
}
\arguments{
\item{bundled_ggplot}{The bundled ggplot from the bundle_ggplot object}
}
\description{
A convenience function so that you can create a file with dimensions as specified in the bundled ggplot object
}
