% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit documents.R
\name{format.round}
\alias{format.round}
\title{Format rounded numbers}
\usage{
\method{format}{round}(x, digits = 1, leading_zero = TRUE, round_half_up = TRUE)
}
\arguments{
\item{x}{The number to be rounded}

\item{digits}{The number of digits to round to}

\item{leading_zero}{TRUE/FALSE, if enabled (default=TRUE) allows the leading zero in a decimal (e.g. the zero in 0.23) when a number is between -1 and 1}

\item{round_half_up}{TRUE/FALSE, if enabled (default=TRUE) uses half-up rounding instead of R's default round to even}
}
\description{
A text convenience function for more readable rounded numbers. Allows for rounding to specified digits (including lagging 0's), keeping leading zeros, and uses half-up rounding
}
