% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit documents.R
\name{format.text.percent}
\alias{format.text.percent}
\title{Formats CIs and point estimates for text output}
\usage{
\method{format}{text.percent}(x, ...)
}
\arguments{
\item{...}{Specify any arguments to be passed to format.round}

\item{point.estimate}{The point estimate of the statistic}

\item{CI.lb}{The confidence interval lower bound of the statistic}

\item{CI.ub}{The confidence interval upper bound of the statistic}

\item{format.percent}{Specifies whether to format the outputs as percentages}

\item{alpha}{The alpha level. E.g. alpha = .05 will show in text as "95\% CI"}

\item{CI.prefix}{= TRUE TRUE/FALSE Enables or disables the text about the alpha bound}

\item{CI.sep}{The text separator between the confidence bounds}

\item{CI.bracket=c("[", "]")}{The brackets for the confidence bounds}
}
\description{
Convenience function that outputs a number and CI bounds as a well formatted text string
}
