% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit documents.R
\name{format.text.CI}
\alias{format.text.CI}
\title{Format a number as a percent}
\usage{
\method{format}{text.CI}(
  point.estimate,
  CI.lb,
  CI.ub,
  ...,
  format.percent = FALSE,
  alpha = 0.05,
  CI.prefix = TRUE,
  CI.sep = " - ",
  CI.bracket = c("[", "]")
)
}
\arguments{
\item{...}{Specify any arguments to be passed to format.round}

\item{x}{The number to be formatted as a percent}
}
\description{
Convenience function which multiplies the number by 100, rounds it, and adds a \% sign
}
