% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit documents.R
\name{knit_docx}
\alias{knit_docx}
\title{Knit placeholders to/from docx files}
\usage{
knit_docx(
  template_docx_file = NA,
  template_drive_ID = NA,
  knitted_docx_file = NA,
  knitted_docx_google_ID = NA,
  placeholder_generation_code_file = NA
)
}
\arguments{
\item{template_docx_file}{The input .docx file containing the placeholders (in this case text in curly brackets{})}

\item{template_drive_ID}{The Google Drive ID of the template document containing the placeholders (in this case text in curly brackets{}). May be either a .docx or a Google Doc.}

\item{knitted_docx_file}{The output .docx file that knits together the figures/text/stats from the code}

\item{knitted_docx_google_ID}{The Google Drive ID of the output knit .docx file that will be overwritten on the knit. Note that this must point to an EXISTING .docx on google drive (can be a blank file uploaded for the purpose)}

\item{placeholder_generation_code_file}{(optional) If NA, the code will extract palceholder objects from the current global environment. Alternatively, this can be pointed to an existing .R script to run, runs the code, and will extract the placeholders from that run environment}
}
\value{
Nothing
}
\description{
This function knits together an output .docx document from a template .docx document with placeholders and analysis code
}
\details{
A package to conveniently enable reproducible workflows for manuscripts
}
