% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit documents.R
\name{preview_bundled_ggplot}
\alias{preview_bundled_ggplot}
\title{Previews bundled ggplot object as it will appear in print}
\usage{
preview_bundled_ggplot(bundled_ggplot)
}
\arguments{
\item{bundled_ggplot}{The bundled ggplot from the bundle_ggplot object}
}
\description{
A convenience function so that you can preview what the bundle_ggplot object will look like in the specified dimensions
}
